+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [feat! (#32): implement `export` modifier](https://github.com/amp-lang/amp/issues/32)
 - [feat (#29): implement `import` statements](https://github.com/amp-lang/amp/issues/29)
 - [feat (#35): implement `namespace`s](https://github.com/amp-lang/amp/issues/35)
+- [feat (#37): `as` conversion operator](https://github.com/amp-lang/amp/issues/37)
 
 ### Library
 - feat: implement standard library, I guess